Fix problems with filesize > 4GiB.
Problem with integer overflow. Problem with using portable types that
contain only 32 bits on 32-bit systems.

+++ src/fileops.h
@@ -56,7 +56,7 @@ void change_group(void);
 /* Returns new value for save_msg flag. */
 int put_files_from_register(FileView *view, int name, int force_move);
 void clone_file(FileView* view);
-off_t calc_dirsize(const char *path);
+unsigned long long calc_dirsize(const char *path);
 
 #endif
